dima_6_stark_mode_chat/requirements.txt
openai
python-dotenv
speechrecognition
pyttsx3
